chore: upgrade to next.js v16
