+++ packages/core/src/workers/WorkerPool.ts
@@ -2,9 +2,9 @@
  * Worker thread pool for parallel processing
  */
 
-import { Worker } from 'worker_threads';
-import { cpus } from 'os';
-import { join } from 'path';
+import { Worker } from 'node:worker_threads';
+import { cpus } from 'node:os';
+import { join } from 'node:path';
 import type { WorkerTask, WorkerResult, WorkerPoolConfig, WorkerPoolStats } from './types';
 
 /**
