--- site/.eslintrc.json
@@ -1,4 +1,0 @@
-{
-  "extends": "next/core-web-vitals"
-}
-
