+++ site/package.json
@@ -4,35 +4,39 @@
   "private": true,
   "scripts": {
     "dev": "next dev",
-    "build": "npm run copy-openredaction && npm run docs:export && next build",
+    "build": "npm run copy-openredaction && npm run docs:export && next build --turbopack",
     "copy-openredaction": "mkdir -p public/lib && cp ../packages/core/dist/index.js public/lib/openredaction.js",
     "start": "next start",
-    "lint": "next lint",
+    "lint": "eslint .",
     "docs:export": "cd docs && npm ci && npm run build && rm -rf ../public/docs && mkdir -p ../public && cp -R out ../public/docs"
   },
   "dependencies": {
     "botid": "^1.5.10",
     "clsx": "^2.1.0",
     "lucide-react": "^0.344.0",
-    "next": "^14.2.0",
+    "next": "16.2.9",
     "nodemailer": "^7.0.12",
-    "openredaction": "file:../packages/core",
-    "react": "^18.3.0",
-    "react-dom": "^18.3.0",
+    "openredaction": "workspace:*",
+    "react": "19.2.7",
+    "react-dom": "19.2.7",
     "resend": "^6.4.2",
     "tailwind-merge": "^2.2.1"
   },
   "devDependencies": {
     "@types/node": "^20.11.0",
     "@types/nodemailer": "^7.0.4",
-    "@types/react": "^18.2.0",
-    "@types/react-dom": "^18.2.0",
+    "@types/react": "19.2.17",
+    "@types/react-dom": "19.2.3",
     "autoprefixer": "^10.4.17",
-    "eslint": "^8.57.0",
-    "eslint-config-next": "^14.2.0",
+    "eslint": "^9",
+    "eslint-config-next": "16.2.9",
     "postcss": "^8.4.33",
     "tailwindcss": "^3.4.1",
     "tsx": "^4.21.0",
     "typescript": "^5.4.0"
+  },
+  "overrides": {
+    "@types/react": "19.2.17",
+    "@types/react-dom": "19.2.3"
   }
 }
